fix i16

+++ test/service.thrift
@@ -1,3 +1,3 @@
 service TestService {
-  i32 doTest(1:i32 id, 2:string str);
+  i32 doTest(1:i16 id, 2:string str);
 }
